connected the Donate button dialog
