+++ ClientApp/src/components/Home.js
@@ -70,9 +70,6 @@ export class Home extends Component {
           <li>
             <a href='/u/settings'>Account Settings</a>
           </li>
-          <li>
-            <a href='/u/donate-modal'>Donate-Modal</a>
-          </li>
         </ul>
       </div>
     );
