+++ ClientApp/src/components/user/Settings.jsx
@@ -4,8 +4,8 @@
  * @author SBD
  */
 
-import React, { useState } from "react";
-import {useNavigate} from "react-router-dom"
+import React, { useState, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 // import { createTheme, ThemeProvider, useTheme } from "@mui/material/styles";
 import Accordion from "@mui/material/Accordion";
 import AccordionSummary from "@mui/material/AccordionSummary";
@@ -23,22 +23,33 @@ import {
   DialogContentText,
   DialogActions
 } from "@mui/material";
+import AccountBalanceIcon from "@mui/icons-material/AccountBalance";
+import SettingsIcon from "@mui/icons-material/Settings";
+import ManageAccountsIcon from "@mui/icons-material/ManageAccounts";
+import AccountSettings from "./AccountSettings";
+import PersonalInformation from "./PersonalInformation"
+import PaymentMethods from "./PaymentMethods"
 
 export const Settings = () => {
+  const [tab, setTab] = useState(0);
   const [newEmail, setNewEmail] = useState("");
   const [newPassword, setNewPassword] = useState("");
   const [openDialog, setOpenDialog] = useState(false);
-  const navigate = useNavigate()
+  const navigate = useNavigate();
   const user = UserInfo();
 
-const StyledButton = styled(Button)(() => ({
-  borderRadius: "20px",
-  backgroundColor: "#B5E3BB",
-  color: "black",
-  fontSize: 12,
-  width: "85%",
-  margin: "0 auto"
-}));
+  useEffect(() => {
+    setTab(0);
+  }, []);
+
+  const StyledButton = styled(Button)(() => ({
+    borderRadius: "20px",
+    backgroundColor: "#B5E3BB",
+    color: "black",
+    fontSize: 12,
+    width: "85%",
+    margin: "0 auto"
+  }));
 
   //added this to handle passworda and email
   const handleUpdate = () => {
@@ -58,9 +69,9 @@ const StyledButton = styled(Button)(() => ({
     axios
       .post(url, payload)
       .then((response) => {
-        navigate("/u/home")
-        setNewEmail("")
-        setNewPassword("")
+        navigate("/u/home");
+        setNewEmail("");
+        setNewPassword("");
         console.log(response.data);
         // show success message to user
       })
@@ -86,9 +97,9 @@ const StyledButton = styled(Button)(() => ({
         userPassword: user.userPassword
       })
       .then((response) => {
-        navigate("/")
-        setNewEmail("")
-        setNewPassword("")
+        navigate("/");
+        setNewEmail("");
+        setNewPassword("");
         console.log(response.data);
         // show success message to user and redirect to login page
         //alert("Your account has been deleted.");
@@ -110,7 +121,6 @@ const StyledButton = styled(Button)(() => ({
       alert("Incorrect password. Please try again.");
     }
   };
-  
 
   return (
     <>
@@ -140,23 +150,52 @@ const StyledButton = styled(Button)(() => ({
                 background: "white"
               }}>
               <img
-                  src='/fund.png'
-                  alt='Avatar'
-                 justify='center'
-                  style={{ width: "100%", height: "100%" }}
-                /> 
+                src='/fund.png'
+                alt='Avatar'
+                justify='center'
+                style={{ width: "100%", height: "100%" }}
+              />
             </Stack>
             <Stack mt={8} borderTop={"1px solid black"} textAlign={"center"}>
               <Stack pt={2} pb={2}>
-                {" "}
-                Personal Information{" "}
+                <Button
+                  onClick={() => setTab(2)}
+                  variant='contained'
+                  sx={{
+                    height: 40,
+                    width: 300,
+                    backgroundColor: "lightgrey",
+                    color: "black"
+                  }}>
+                  <ManageAccountsIcon /> Personal Information{" "}
+                </Button>
               </Stack>
               <Stack pt={2} pb={2}>
-                {" "}
-                Payment Methods{" "}
+                <Button
+                  variant='contained'
+                  onClick={() => setTab(1)}
+                  sx={{
+                    height: 40,
+                    width: 300,
+                    backgroundColor: "lightgrey",
+                    color: "black"
+                  }}>
+                  <AccountBalanceIcon /> Payment Methods{" "}
+                </Button>
               </Stack>
               <Stack mb={8} pt={2} pb={2}>
-                Account Settings{" "}
+                <Button
+                  onClick={() => setTab(0)}
+                  variant='contained'
+                  sx={{
+                    height: 40,
+                    width: 300,
+                    backgroundColor: "lightgrey",
+                    color: "black"
+                  }}>
+                  <SettingsIcon />
+                  Account Settings{" "}
+                </Button>
               </Stack>
 
               <Stack mt={15}>
@@ -165,130 +204,19 @@ const StyledButton = styled(Button)(() => ({
             </Stack>
           </Stack>
         </Stack>
-        <Stack
-          position='relative'
-          backgroundColor={"lightgrey"}
-          borderRadius='20px'
-          border='1px solid black'
-          pt={8}
-          width='70%'
-          direction='column'
-          textAlign='left'>
-          <Accordion>
-            <AccordionSummary
-              expandIcon={<ExpandMoreIcon />}
-              aria-controls='panel1a-content'
-              id='panel1a-header'>
-              <Typography>
-                <b>Update Email</b>
-              </Typography>
-            </AccordionSummary>
-            <AccordionDetails>
-              <Typography>Enter your new email:</Typography>
-              <TextField
-                fullWidth
-                lablel='Email'
-                id='newEmail'
-                placeholder='Enter new email:'
-                value={newEmail}
-                onChange={(e) => setNewEmail(e.target.value)}
-              />
-            </AccordionDetails>
-            <AccordionDetails>
-              <Typography>Re-Enter your new email:</Typography>
-              <TextField
-                fullWidth
-                lablel='New Email'
-                id='newEmail'
-                placeholder='Re-Enter new email:'
-              />
-            </AccordionDetails>
-          </Accordion>
-          <Accordion>
-            <AccordionSummary
-              expandIcon={<ExpandMoreIcon />}
-              aria-controls='panel1a-content'
-              id='panel1a-header'>
-              <Typography>
-                <b>Update Password</b>
-              </Typography>
-            </AccordionSummary>
-            <AccordionDetails>
-              <Typography>Enter your new password</Typography>
-              <TextField
-                fullWidth
-                lablel='New Email'
-                id='newEmail'
-                placeholder='Enter new password:'
-                type='password'
-                value={newPassword}
-                onChange={(e) => setNewPassword(e.target.value)}
-              />
-            </AccordionDetails>
-            <AccordionDetails>
-              <Typography sx={{ width: "60%", flexShrink: 0 }}>
-                Re-Enter your new password
-              </Typography>
-              <TextField
-                fullWidth
-                lablel='New Password'
-                id='newPassword'
-                placeholder='Re-Enter new password:'
-                type='password'
-              />
-            </AccordionDetails>
-          </Accordion>
-
-          <Stack
-            position='absolute'
-            width='100%'
-            bottom={"10px"}
-            direction='row'
-            columnGap={3}
-            justifyContent='center'>
-            <Button
-              variant='contained'
-              color='error'
-              sx={{ height: 40, borderRadius: "20px", width: 200 }}
-              onClick={handleDelete}>
-              Delete Account
-            </Button>
-          </Stack>
-          <Dialog
-            open={openDialog}
-            onClose={handleDialogClose}
-            aria-labelledby='alert-dialog-title'
-            aria-describedby='alert-dialog-description'>
-            <DialogTitle id='alert-dialog-title'>
-              {"Delete Account"}
-            </DialogTitle>
-            <DialogContent>
-              <DialogContentText id='alert-dialog-description'>
-                Are you sure you want to delete your account?
-              </DialogContentText>
-            </DialogContent>
-            <DialogActions>
-              <StyledButton onClick={handleDialogClose}>Cancel</StyledButton>
-              <StyledButton onClick={handleAccountDelete} autoFocus>
-                Delete
-              </StyledButton>
-            </DialogActions>
-          </Dialog>
-          <Stack sx={{ position: "absolute", bottom: "10px", right: "10px" }}>
-            <Button
-              variant='contained'
-              sx={{
-                height: 40,
-                width: 200,
-                backgroundColor: "#B5E3BB",
-                color: "black",
-                borderRadius: "20px"
-              }}
-              onClick={handleUpdate}>
-              Update
-            </Button>
-          </Stack>
-        </Stack>
+       {tab===0 && <AccountSettings
+          setNewEmail={setNewEmail}
+          newEmail={newEmail}
+          setNewPassword={setNewPassword}
+          newPassword={newPassword}
+          handleAccountDelete={handleAccountDelete}
+          handleUpdate={handleUpdate}
+          handleDelete={handleDelete}
+          handleDialogClose={handleDialogClose}
+          openDialog={openDialog}
+        />}
+        {tab===2 && <PersonalInformation />}
+        {tab===1 && <PaymentMethods/>}
       </Stack>
     </>
   );
