+++ package.json
@@ -3,6 +3,7 @@
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
     "@hookform/resolvers": "^2.9.11",
+    "@mui/icons-material": "^5.11.16",
     "@mui/lab": "^5.0.0-alpha.122",
     "@mui/material": "^5.11.14",
     "@mui/styled-engine-sc": "^5.11.11",
